refactor: move get_buffer() to v8serialize._pycompat.typing
It's not specific to jsbuffers and we use it elsewhere now. We also no
longer export it, I don't think it should be part of our API.
